Make ActivityPaymentMethodCallback an inner class of AddPaymentMethodActivity
